Only allow .jsx extension for files containing JSX
In order to ensure the build is performant, vite is configured to
only run the JSX transform on files with a .jsx extension.

Ensure that only files containing JSX have a .jsx extension to avoid
unnecessary processing.
