+++ packages/components/src/components/DotSpinner/DotSpinner.stories.js
@@ -1,5 +1,5 @@
 /*
-Copyright 2019-2023 The Tekton Authors
+Copyright 2021-2024 The Tekton Authors
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
@@ -11,13 +11,11 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import Graph from './Graph';
-
-import graph from './examples/graph.json';
+import DotSpinner from './DotSpinner';
 
 export default {
-  component: Graph,
-  title: 'legacy/Graph'
+  component: DotSpinner,
+  title: 'DotSpinner'
 };
 
-export const Default = { args: { graph } };
+export const Default = {};
