+++ packages/components/src/components/Spinner/Spinner.stories.js
@@ -1,5 +1,5 @@
 /*
-Copyright 2020-2023 The Tekton Authors
+Copyright 2021-2024 The Tekton Authors
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
@@ -11,11 +11,11 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import LoadingShell from './LoadingShell';
+import Spinner from './Spinner';
 
 export default {
-  component: LoadingShell,
-  title: 'LoadingShell'
+  component: Spinner,
+  title: 'Spinner'
 };
 
 export const Default = {};
