+++ packages/components/src/components/TextInput/TextInput.stories.js
@@ -1,5 +1,5 @@
 /*
-Copyright 2021-2023 The Tekton Authors
+Copyright 2019-2024 The Tekton Authors
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
@@ -13,27 +13,24 @@ limitations under the License.
 
 import { action } from '@storybook/addon-actions';
 
-import DeleteModal from './DeleteModal';
+import TextInput from './TextInput';
 
 export default {
-  component: DeleteModal,
-  title: 'DeleteModal'
+  args: {
+    id: 'text-input-id',
+    onChange: action('onChange'),
+    onClick: action('onClick')
+  },
+  component: TextInput,
+  title: 'TextInput'
 };
 
 export const Default = {
   args: {
-    kind: 'Pipelines',
-    onClose: action('onClose'),
-    onSubmit: action('onSubmit'),
-    resources: [
-      {
-        metadata: {
-          name: 'my-pipeline',
-          namespace: 'my-namespace',
-          uid: '700c9915-65f0-4309-b7e0-54d2e4dc8bea'
-        }
-      }
-    ],
-    showNamespace: false
+    labelText: 'foo',
+    helperText: 'this is a description of input foo',
+    placeholder: 'bar'
   }
 };
+
+export const Loading = { args: { loading: true } };
