+++ packages/graph/src/components/Graph.stories.js
@@ -1,5 +1,5 @@
 /*
-Copyright 2021-2023 The Tekton Authors
+Copyright 2019-2024 The Tekton Authors
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
@@ -11,11 +11,13 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import DotSpinner from './DotSpinner';
+import Graph from './Graph';
+
+import graph from './examples/graph.json';
 
 export default {
-  component: DotSpinner,
-  title: 'DotSpinner'
+  component: Graph,
+  title: 'legacy/Graph'
 };
 
-export const Default = {};
+export const Default = { args: { graph } };
