+++ src/containers/LoadingShell/LoadingShell.stories.js
@@ -1,5 +1,5 @@
 /*
-Copyright 2019-2023 The Tekton Authors
+Copyright 2020-2024 The Tekton Authors
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
@@ -11,26 +11,11 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import { action } from '@storybook/addon-actions';
-
-import TextInput from './TextInput';
+import LoadingShell from './LoadingShell';
 
 export default {
-  args: {
-    id: 'text-input-id',
-    onChange: action('onChange'),
-    onClick: action('onClick')
-  },
-  component: TextInput,
-  title: 'TextInput'
+  component: LoadingShell,
+  title: 'LoadingShell'
 };
 
-export const Default = {
-  args: {
-    labelText: 'foo',
-    helperText: 'this is a description of input foo',
-    placeholder: 'bar'
-  }
-};
-
-export const Loading = { args: { loading: true } };
+export const Default = {};
